requirements.txt
flask
pdfplumber
pandas
openpyxl
sqlalchemy
python-dotenv
gunicorn